+++ InplaceX-android/app/src/main/java/com/mirkori/inplacex/ui/screens/home/HomeScreenState.kt
@@ -2,6 +2,7 @@ package com.mirkori.inplacex.ui.screens.home
 
 enum class HomeScreenState {
     ROOT,
+    PVP_MODES,
     PVE_GAME,
     PVP_GAME
 }